packages - fix subversion version checks
Use "Last Changed Rev:" rather than "Revision:" as we want the last revision for the URL/branch/tag
rather than the last revision for the repository.

Don't just compare commits if the commit is HEAD (used for svn repos) as it's not a static commit but
references the latest commit.