vice - switch to v3.5 branch
There are audio issues on the RPI with the current HEAD version. But also we have had to update this
scriptmodule a fair amount due to upstream changes, and it would be better now to stick to the latest
stable version. Previously there were less frequent "stable" releases also.

The v3.5 version seems to work without issue.