dxx-rebirth - temporarily roll back code due to a configure failure on Raspbian Buster
The commit after has a change to use pkg-config for locating physfs and it seems to fail on Debian/Raspbian Buster.
Will debug properly and feed upstream, but for now am rolling back to the commit before.

Unfortunately the dxx-rebirth repository has no recent releases at all, as it would be preferable to lock
to the last stable release, but the last tag is from 2018 and there are no version branches etc.