drastic - rework module to use setBackend
Previously the module used XINIT: to force X on kms systems (rpi4), but now with the new backend
configuration, we can use the same emulator.cfg launch parameters, but default it to X11 for kms.

Also remove the X11 launch code, as runcommand will take care of this now.

Moved the logic around also, to avoid install related configuration work on removal.