mupen64plus - use apt-cache to check cmake version rather than hasPackage (dpkg)
The _get_repos_mupen64plus function is used for update checks, and it's called before depends_mupen64plus in this case.
If cmake was removed after installing, the hasPackage would fail and an older commit for the GLideN64 repo would
be used, which would cause an update even if not required.

This is unlikely to happen for end users, but the builder module removes dependencies after installing (so missing
dependencies can be identified). This caused mupen64plus binaries to be rebuilt when there had been no changes.

This check is only needed for older pre-buster distros with cmake older than 3.9 so will be removed in the future.