packages - use last commit date rather than file date for checking module modifications
rp_hasNewerModule will force a rebuild if the module was updated since it was installed. This was added to
handle cases where we update patches or build parameters in module. Using the last commit date is more reliable
as module dates can be changed when rebasing etc. This resolves some issues with the builder module rebuilding
binaries when not needed, due to branch switches / rebases.