helpers / download - added --no-buffer when dest is stdout to avoid errors when piping
Curl can throw an error "(23) Failed writing body" if piping the output to grep, or something that closes the stream
early due to the way curl buffers. Adding --no-buffer when output is stdout (dest of "-") fixes this. This error
didn't cause any failures but _get_branch_mame could cause this error to be logged.

Also switched "-s" to "--silent". We have a lot of curl arguments and it's more intuitive to use the descriptive
names.