packages - allow custom binary installs when we don't host binaries
Custom binary URLs provided by __binary_url_* scriptmodule functions were ignored for platforms
we don't host general binary builds for. This meant that lr-duckstation wouldn't show binary
install options on aarch64 / x86_64.

Moved the rp_hasBinaries check logic to rp_getBinaryUrl, so we only set a URL by default when
we have binaries for that system. This is overridden as needed by install_bin_* / __binary_url_*
scriptmodule functions. The rp_hasBinary function will check for a binary if the URL is set, so
we allow overriding of our own binaries, as well as having the ability to install 3rd party binaries
for systems we don't host binaries for.