lr-parallel-n64 - switch to our own fork temporarily to fix segfault on exit on arm
Had to revert a couple of patches as upstream did another workaround for the sefault on exit
which doesn't work for me. As upstream isn't too active, it's quicker/safer to fix in a fork and switch back
later than include a couple of patches which could break at some point.