atari800 - enable configuration of sdl1 backends for module
Split out the addEmulator calls to a separate function as we need to change them based on the configured backend.

Rework the parameters in the new function, to handle launching via X as well as on FB via kms.

Default to "dispmanx" backend for RPI4 on fkms, as it should be fastest. It will however look blurry as it will upscale
from the native atari800 resolution and the scaling uses a filter. However this is still preferable from software
upscaling to native resolution which was the previous default for kms. RPI4 Users can also switch to X11 backend, which
will allow for opengl acceleration for scaling and supports video mode switching from runcommand.sh.