Dosbox Staging - move to tag from now on instead of master/main
+ automatically get latest tag (thanks @dankcushions)

The main reason to switch from master branch to a fixed version (tag) is all about stability. Heavy changes are to be worked on shortly on DOSBox Staging (scaling method, better shader integration, input handling, etc just to name a few). Therefore the master branch is likely to break at some point.

In order to not overwhelm Github with tickets (or Retropie forum) when build is broken, let's stick to a fixed version (starting with 0.77.0). The fixed/tagged branches are stable and have been tested on various architectures/devices.