packages - work from commit date rather than authored date
This is used for our package information, although it's not used to check for newer versions
as we only check the remote hash of a git repository for updates.

We also used this for checking if we need to update a binary based on the last date of a change to the
module file. However as we were using author date, if a binary was built before the change was accepted, but
more recently than the change was authored, it wouldn't trigger a binary rebuild.

It makes more sense to use the date the commit was made rather than when it was authored in both cases, as
we want to know when something was last changed in the repository rather than the date of the change.