helpers / gitPullOrClone - fixup / improve persistent repos support
When env __persistent_repos is 1, we do full checkouts of repositories. This allows users with diskspace to spare
to speed up source updates, by avoiding doing a shallow clone everytime.

However, it was not working for cases where a module switched to a newer tag/branch as remote changes were not
fetched first. This is now fixed.

Also fixed errors thrown when calling gitPullOrClone on an existing repository when doing git pull, when HEAD
is detached (for tags) or there is no tracking branch.