runcommand launch dialog fix
This change allows the launch menu dialog to still be shown when launch art is enabled if no image is found. Current behavior is that when art is enabled, the dialog is never shown; therefore if no image is found, nothing is displayed at all.

This line to show the dialog is only called if no image is found in the first place, so checking if art is enabled at this point seems superfluous. We didn't get one, for one reason or another; either because it wasn't there, or because we didn't even look. By eliminating this test, we can fall back to show the dialog whenever an image is not obtained, for WHATEVER reason.