advmame: fix build with newer GCC for older versions
With gcc10 and newer, the `-fno-common` option is the default, causing
issues for global variables with multiple declarations that are missing the `external` initializer.
See https://wiki.gentoo.org/wiki/Project:Toolchain/Gcc_10_porting_notes/fno_common for a summary.

Added a couple of patches to fix building for older `advmame` versions:
 * for `advmame-1.4`, pick https://github.com/amadvance/advancemame/commit/f8688cae05
 * for `advmame-0.94`, created a separate patch

As a minor modification, used the scriptmodule `md_path` variable to calculate the patch folder location.