lr-stella2014 - remove renaming hook
Now we have a new upstream lr-stella we don't want this rename anymore as it will rename the new module.

If a user upgrades from a very old release with the older lr-stella module, they will get upgraded to the new core but can always install lr-stella2014 if they want the older core again.