flycast: add standalone emulator
Add the Flycast standalone emulator, in addition to the libtretro core we already have.
Supports the same systems/platforms and has also RetroAchievements support (automatically built in).

The emulator has an in-game menu, opened by 'Tab' on a keyboard or - by default - with the 'Guide'/'Menu' button on a controller. The emulator uses the SDL Gamecontroller system to auto-configure a gamepad (this is how the 'Guide'/'Menu' buttons are found) and recently has added hotkey/button combinations support.
ocal modifications