ppsspp / lr-ppsspp - switch to v1.12.3 (last tagged version)
Current master has shader issues on videocore causing a black screen, so switching back to last tagged version - v1.12.3.

Historically we have had to fix up a fair number of ppsspp issues on the RPI due to upstream changes so sticking to a stable release seems like a good idea and allows us to test before we update.

The PPSSPP repository is also quite active and re-building ppsspp/lr-ppsspp binaries for every change takes a lot of time, so this also frees up some cpu time on our build system.