helpers - fix sdl1 dependency installing on non RPi platforms.
In the past unsupported scriptmodules for a platform were not loaded, so rp_hasModule would only return true if the module was available for the platform.

As we now load all modules for displaying of information in retropie_setup / retropie_packages output, rp_hasModule function will return true, so we need to use rp_isEnabled instead.

This fixes installing scriptmodules with an sdl1.2-dev dependency on non RPi (eg x86/x11), that previously forced our custom sdl1 and then failed as it's disabled for the target.