helpers - added inputBox function that uses the OSK
If the OSK is installed it will be used for input, otherwise it will fall back to using dialog --inputbox

function takes 3 parameters:

The title of the dialog, the default text and a minimum number of characters to accept.

Adjusted wifi module to use the new inputBox. There is no need to redirect cmd output with 2>&1 >/dev/tty as the inputBox function does this.