helpers / hasPackage - fix package detection on a multiarch system
On a multiarch system both amd64 and i386 versions of a package may be installed.

This would cause both to be returned by dpkg-query on a single line causing the extracting of version and install status to fail.

It now splits the output onto multiple lines and grabs the first line to workaround this.

Fixes #3679