image - use the mke2fs config from the chroot for ext4 filesystem
This ensures the filesystem is created with supported features if images are created on a system with a newer mkfs.ext4.

Remove metadata_csum disabling as it's no longer needed with above changes (and was added for compatibility with Jessie).

Disable 64bit ext4 feature as with the Raspberry Pi OS images.