builder - rework chroot builder RetroPie-Setup cloning
Previously RetroPie-Setup was cloned from the calling RetroPie-Setup scriptmodule on the host side, but this only worked when the UID/GID of the user was 1000:1000 matching the first user in the chroot.

Due to changes in my build system, this required a fix. It now does the clone from within the chroot, so the ownership is correct for chroot pi user.

This commit also fixes up the multiline quoting and removes an unnecessary trailing slash on a mkdir.