Turn on SRM check and autosave feature
When playing a game on RetroPie if there is a sudden disconnection that isn't clean (power cut, pulling out the power cable etc), the .srm save file isn't updated correctly. The "autosave_interval" feature checks if the .srm has been updated and if it has been updated it overwrites the original file. In the case of the user, this would be expected behaviour (if you saved your game the save file would write then, not when you exit Retroarch).

The feature itself isn't a significant strain on resources, playing N64, PS1 cores on a pi4 I didn't notice any difference in frame rates or performance.