lr-flycast-dev: add upstream libretro core
Added the libretro core from upstream flyinghead/flycast, it supercedes the current libretro/flycast repository used for the current `lr-flycast` core.
The new core doesn't work correctly with the legacy VideoCore GLES drivers [1], so it will have to co-exist with the current `lr-flycast` until we drop support for older RaspiOS versions still using the VideoCore drivers.

Due to an internal compiler error, the enw core doesn't build on Buster with the included `gcc8`, but works on `gcc9` and later, which means it can be installed on Debian 11/Ubuntu 20.04 and later.

Tested on RaspiOS _bookworm_ with `vulkan`, `glcore` and `gl` RetroArch video drivers, on a Pi4.

[1] https://github.com/flyinghead/flycast/issues/559