smw: switch to @mmatyas's fork, at version 2.0
The SMW fork from mmatyas/supermariowar is still active and it has the options to use SDL2 instead of SDL1.
Switched the scriptmodule to it with updated dependencies and build instructions.
New version also changed the configuration location, now in a dedicated folder.

Main changes since 1.8

 * Experimental online mode, including an online lobby server
 * Lots and lots and lots of bugfixes (seriously, there's at least 200)
 * Added Web, ARM and (an experimental) Android ports, and improved support for Windows, Linux and OSX
 * Improved music and sound quality
 * Ported all parts of the game to SDL2, for hardware accelerated drawing and improved support for many platforms
 * Major cleanup of the source code and the build system

NOTE: I've updated the license to GPL2, since this is what's declared in the CMake build file.