initfuncs: change `sed` separator for `iniGet`
Use `#` instead of `/` for SED's substitution command separator.

This makes the `iniGet` command work for key values containing `/` (like Dolphin's input configuration files). The `#` character is usually reserved for comment lines in an `.ini` file, so it should be relatively safe to use as separator; `sed` accepts any character besides `*` or newline for separator, so `#` should also be safe here.